+++ apps/backoffice-v2/src/lib/blocks/components/KycBlock/hooks/useKycBlock/useKycBlock.tsx
@@ -114,6 +114,9 @@ export const useKycBlock = ({
                 vendor: string;
               };
             };
+            kycDocumentDetails?: {
+              expiryDate: string;
+            };
             documents: TDocument[];
           };
     }