+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.122
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/ui@0.7.159
+
+## 1.2.121
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.158
+
+## 1.2.120
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.156
+- bump
+- Updated dependencies
+  - @ballerine/config@1.1.44
+  - @ballerine/ui@0.7.158
+
 ## 1.2.119
 
 ### Patch Changes